requirements.txt
matplotlib~=3.10.0
numpy~=2.2.2
pandas~=2.2.3
seaborn~=0.13.2
omegaconf~=2.3.0
scikit-learn~=1.6.1
torch-geometric~=2.6.1
tornado~=6.4.2
mlflow~=2.21.0
mlflow-skinny~=2.21.0
hydra-core~=1.3.2
typing_extensions~=4.12.2